# Optimizes reference count of Span and squashes an errorprone warning