--- a/pom.xml
+++ b/pom.xml
@@ -91,7 +91,7 @@
 
     <!-- Catch common Java mistakes as compile-time errors -->
     <plexus-compiler-javac-errorprone.version>2.8.2</plexus-compiler-javac-errorprone.version>
-    <errorprone.version>2.1.3</errorprone.version>
+    <errorprone.version>2.2.0</errorprone.version>
   </properties>
 
   <name>Zipkin (Parent)</name>