… zipkin-storage/elasticsearch/src/test/java/zipkin2/elasticsearch/ElasticsearchSpanConsumerTest.java
… zipkin2/src/main/java/zipkin2/storage/QueryRequest.java