Based on https://www.youtube.com/watch?time_continue=2118&v=YLTD1GBEfwo
we were a bit noisy for no good reason. Thanks @raphw for noticing

This also squashes an error-prone warning, which while wasn't a bug
wasn't worth explaining in a suppression why it wasn't.