@adriancole 2018-01-19
Here are the benchmark results

Before
```
SpanBenchmarks.buildClientOnlySpan2        thrpt   15   4.222 ± 0.128  ops/us
SpanBenchmarks.buildClientOnlySpan2_clear  thrpt   15   4.303 ± 0.101  ops/us
SpanBenchmarks.buildClientOnlySpan2_clone  thrpt   15   5.754 ± 0.184  ops/us
```

After
```
SpanBenchmarks.buildClientOnlySpan2        thrpt   15   4.288 ± 0.152  ops/us
SpanBenchmarks.buildClientOnlySpan2_clear  thrpt   15   4.517 ± 0.167  ops/us
SpanBenchmarks.buildClientOnlySpan2_clone  thrpt   15   6.144 ± 0.074  ops/us
```