@adriancole 2018-01-24
added this to follow-up in brave https://github.com/openzipkin/brave/pull/589